Zend_Test works with both PHPUnit 3.4 and 3.7+
I know, it is a little bit hacky solution, but it makes it possible to use controller tests in PHPUnit 3.4 as well as any higher version.
I suppose, that ZF1 is in some maintance mode, where the main goal is to have it running on newer SW versions and if someone wants to have superclean code, he can use ZF2 ;)

Inpired by the patch from http://opensource.hqcodeshop.com/Zend%20Framework/PHPUnit%203.7%20-%201.12.1/
